Moved error classes to error utility
